requirements.txt
ipykernel
google-cloud-storage
h5py==2.10.0
keras
matplotlib
notebook
numpy==1.19.2
pandas==1.1.4
seaborn
sklearn
scipy==1.4.1
tensorflow==2.2
